Test both implementations of ChangeLock
