+++ shared/src/test/java/org/odk/collect/shared/locks/BooleanChangeLockTest.kt
@@ -1,0 +1,7 @@
+package org.odk.collect.shared.locks
+
+class BooleanChangeLockTest : ChangeLockTest() {
+    override fun buildSubject(): ChangeLock {
+        return BooleanChangeLock()
+    }
+}
